fix: move state out of DimensionItemBase
